fix conversation warnings

+++ src/lv_misc/lv_mem.c
@@ -107,7 +107,7 @@ void lv_mem_init(void)
  * @param size size of the memory to allocate in bytes
  * @return pointer to the allocated memory
  */
-void * lv_mem_alloc(uint32_t size)
+void * lv_mem_alloc(size_t size)
 {
     if(size == 0) {
         return &zero_mem;
@@ -222,7 +222,7 @@ void lv_mem_free(const void * data)
 
 #if LV_ENABLE_GC == 0
 
-void * lv_mem_realloc(void * data_p, uint32_t new_size)
+void * lv_mem_realloc(void * data_p, size_t new_size)
 {
     /*data_p could be previously freed pointer (in this case it is invalid)*/
     if(data_p != NULL) {
